requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
flake8==6.1.0
geojson==2.5.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
PySocks==1.7.1
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
soupsieve==2.5
text-unidecode==1.3
urllib3==2.0.4
vk-api==11.9.9
